He editado alguna cosa del dame_Onn y he hecho el test (mirarlo de todas formas por si no os convence)

+++ Codigos_en_C/funciones_red.h
@@ -75,7 +75,7 @@ int indice_lado(int *plaquetas, int *aristas, int arista_actual);
 
 void indices_esquinas(int *aristas, int *plaquetas, int nodo, int plano, int arista_1, int arista_2, int N_esquina, int *inidce_S1, int *indicce_S2, int *indice_Up);
 
-void dame_O_n(int *aristas, int *plaquetas, int *O, int n, double tabla_promedio[4][4][2], double tabla_spin[5]);
+void dame_O_n(int *aristas, int *plaquetas, double *O, int n, double tabla_promedio[4][4][2], double tabla_spin[5]);
 
 double un_loop_O_x(int Nodo_inicial, int *aristas, int *plaquetas, int n, double tabla_promedio[4][4][2], double tabla_spin[5]);
